Get keypadActive and setKeypadActive in wrapper component
