+++ packages/math-input/src/components/__tests__/integration.test.tsx
@@ -58,14 +58,12 @@ function InputWithContext({keypadConfiguration}) {
 function KeypadWithContext() {
     return (
         <KeypadContext.Consumer>
-            {({setKeypadElement, keypadActive, setKeypadActive}) => {
+            {({setKeypadElement}) => {
                 return (
                     <MobileKeypad
                         onElementMounted={setKeypadElement}
                         onDismiss={() => {}}
                         onAnalyticsEvent={async () => {}}
-                        keypadActive={keypadActive}
-                        setKeypadActive={setKeypadActive}
                     />
                 );
             }}
