+++ packages/perseus/src/__tests__/article-renderer.test.tsx
@@ -22,14 +22,12 @@ import type {APIOptions} from "../types";
 function KeypadWithContext() {
     return (
         <KeypadContext.Consumer>
-            {({setKeypadElement, keypadActive, setKeypadActive}) => {
+            {({setKeypadElement}) => {
                 return (
                     <MobileKeypad
                         onElementMounted={setKeypadElement}
                         onDismiss={() => {}}
                         onAnalyticsEvent={async () => {}}
-                        keypadActive={keypadActive}
-                        setKeypadActive={setKeypadActive}
                     />
                 );
             }}
